clean pkg setup
