--- projects/ffe/config-test.ini
@@ -1,75 +1,0 @@
-[io]
-outdir: "out"
-interval: 1       #output frequency in units of simulation steps for analysis files
-full_interval: -1 #output frequency to write full simulation snapshots
-restart:  100     #frequency to write restart files (these overwrite previous files)
-stride: 1         #output reduce factor; NxMesh/stride must be int
-laprestart: -1    #restart switch (-1 no restart; 0 automatic; X lap to restart)
-
-
-#simulation parameters
-[simulation]
-cfl: 1.0         #time step in units of CFL
-Nt: 5            #number of simulation steps
-
-
-[problem]
-delgam:  2.0e-4    #temperature
-temperature_ratio: 1.0  #T_i/T_e
-
-me: -1.0         #electron mass-to-charge
-mi:  1.0         #ion mass-to-charge
-
-sigma: 10.0      # magnetization number (omega_ce/omega_pe)^2, including gamma for inertia
-
-external_fields: True
-sigma_ext: 0.1   # strength of external magnetization, (omega_ce/omega_pe)^2, including gamma
-
-gamma: 1.1
-
-dstripe:  10.0    # current sheet thickness (in skin depths)
-dvstripe:  1.0    # width of the triggering point (in skin depths)
-nstripe:  3.0     # current sheet overdensity (\eta parameter)
-
-trigger: False    # trigger collapse by removing plasma pressure support in the middle of the domain
-trigger_field: 0.1 # alternative E_z field that penetrates current sheet to trigger collapse
-
-btheta: 0.0
-bphi:   90.0
-bxinit: 0.0      #in units of B0, value of the field across the current sheet
-
-
-min_mode: 1
-max_mode: 2
-drive_ampl: 1.0
-
-
-
-#spatial grid parameters 
-[grid]
-Nx:     8
-Ny:     8
-Nz:     1
-NxMesh: 10
-NyMesh: 10
-NzMesh: 1
-
-c_omp: 5
-
-periodicx: True
-periodicy: True
-periodicz: True
-
-outflowbc: True
-
-#individual velocity mesh parameters
-[vmesh]
-dvx: 1
-vxmin: 0
-vxmax: 1
-
-[particles]
-ppc: 1          #particle per cell per species
-Nspecies: 2     #number of species (typically 2)
-npasses: 0     #number of current filter passes
-
