+++ projects/ffe/pkg_small.ini
@@ -1,0 +1,59 @@
+[io]
+outdir: "test-ffe_2pkg_a1"
+interval: 10      #output frequency in units of simulation steps for analysis files
+full_interval: -1 #output frequency to write full simulation snapshots
+restart:  1000    #frequency to write restart files (these overwrite previous files)
+stride: 1         #output reduce factor; NxMesh/stride must be int
+laprestart: 0     #restart switch (-1 no restart; 0 automatic; X lap to restart)
+
+
+#simulation parameters
+[simulation]
+cfl: 0.45        #time step in units of CFL
+Nt: 1000         #number of steps
+Nspecies: 2      #number of species (typically 2)
+
+me: -1.0        #electron mass-to-charge
+mi: +1.0        #ion mass-to-charge
+
+npasses: 4              #number of current filter passes
+
+[problem]
+
+#prtcl parameters
+delgam:  1.0e-5         #temperature
+temperature_ratio: 1.0  #T_i/T_e
+sigma: 40.0             # magnetization number (omega_ce/omega_pe)^2, including gamma for inertia
+
+#packet initialization
+zeta: 1.0  #perturbation amplitude
+ell: 10.0  #packet width
+impact_param: 0.0 
+two_wave: True         # second wave packet in opposite direction
+two_wave_reversed: False
+
+
+#spatial grid parameters 
+[grid]
+Nx: 4
+Ny: 4
+Nz: 8
+NxMesh: 10
+NyMesh: 10
+NzMesh: 10
+
+dx: 1.0 
+dy: 1.0
+dz: 1.0
+
+#individual velocity mesh parameters
+[vmesh]
+Nvx: 1
+Nvy: 1
+Nvz: 1
+
+[particles]
+ppc: 1      #particle per cell per species
+c_omp: 1
+
+n_test_prtcls: 100000
