--- projects/ffe/reconn.ini
@@ -1,65 +1,0 @@
-[io]
-outdir: "outv4t2"
-interval: 10     #output frequency in units of simulation steps for analysis files
-full_interval: -1 #output frequency to write full simulation snapshots
-restart:  100     #frequency to write restart files (these overwrite previous files)
-stride: 1         #output reduce factor; NxMesh/stride must be int
-laprestart: -1    #restart switch (-1 no restart; 0 automatic; X lap to restart)
-
-
-#simulation parameters
-[simulation]
-cfl: 0.5            #time step in units of CFL
-Nt: 1000            #number of simulation steps
-
-
-[problem]
-delgam:  2.0e-4    #temperature
-temperature_ratio: 1.0  #T_i/T_e
-
-me: -1.0         #electron mass-to-charge
-mi:  1.0         #ion mass-to-charge
-
-
-gamma: 1.1       # prtcl bulk flow in background 
-sigma: 10.0      # magnetization number (omega_ce/omega_pe)^2, including gamma for inertia
-sigma_ext: 1.0   # strength of external magnetization including gamma (in units of initial sigma)
-
-sheet_thickness:  0.0    # current sheet thickness (in skin depths)
-pinch_width:     10.0    # width of the triggering point (in skin depths)
-sheet_density:    3.0    # current sheet overdensity (\eta parameter)
-
-trigger: True       # trigger collapse by removing plasma pressure support in the middle of the domain
-trigger_field:  0.0 # alternative E_z field that penetrates current sheet to trigger collapse
-
-periodicx: True
-periodicy: True
-periodicz: True
-
-outflowbc: True
-
-
-
-#spatial grid parameters 
-[grid]
-Nx:     8
-Ny:     8
-Nz:     2
-NxMesh: 10
-NyMesh: 10
-NzMesh: 10
-
-c_omp: 1
-
-
-#individual velocity mesh parameters
-[vmesh]
-dvx: 1
-vxmin: 0
-vxmax: 1
-
-[particles]
-ppc: 1          #particle per cell per species
-Nspecies: 2     #number of species (typically 2)
-npasses: 0     #number of current filter passes
-
